requirements.txt
requests
python-nmap
dnspython
python-whois
beautifulsoup4
pandas
matplotlib
jinja2

